@@ -1,20 +1,21 @@
 source 'http://rubygems.org'
-gem 'rails', '3.2.21'
+
+ruby '2.2.0'
+gem 'rails', '4.0.13'
+
+## 06/15
+gem 'rails-observers'
 
 ###19/03
-
-gem "paranoia", "~> 1.0"
-
-###
-
+gem 'bullet'
+gem "paranoia", "~> 2.0"
 gem 'unicorn', group: :production
-
 gem 'thin', group: :development
 gem 'rb-readline', group: :development
 gem 'quiet_assets', group: :development
 gem 'sqlite3', group: :test
 
-gem "delocalize", '0.3.2'
+gem "delocalize"
 gem 'execjs'
 gem 'therubyracer'
 
@@ -25,26 +26,29 @@ gem 'mysql2'
 gem 'will_paginate', '3.0.7'
 gem 'will_paginate-bootstrap', '0.2.5'
 
-gem 'devise'
+gem 'devise', '3.5.1'
 gem 'devise-encryptable'
-gem 'simple_form', '2.0.3'
-gem 'inherited_resources', '1.3.0'
+gem 'simple_form', '3.1.0'
+gem 'inherited_resources', '1.6.0'
 gem 'slim', '1.2.1'
 gem 'country_select', '0.0.1'
-gem "liquid", "2.6.1"
-gem 'carrierwave', "0.6.2"
+gem "liquid", "3.0.3"
+
+gem 'carrierwave', "0.7.0"
+gem 'mini_magick', '4.2.7'
+
 gem 'pry', "0.10.1"
 gem 'pry-nav', "0.2.4"
 gem 'pry-stack_explorer', "0.4.9.1"
-gem 'mini_magick', '3.4'
+
 gem "correios-frete", "1.6.1"
-gem "ruby-plesk",         git:  "git://github.com/bperucchi/ruby-plesk.git"
-gem "moip-ruby", "0.1.2", git: "git://github.com/bperucchi/moip-ruby.git"
+# gem "ruby-plesk",         git:  "git://github.com/bperucchi/ruby-plesk.git"
+gem "moip-ruby", "0.1.3", git: "git://github.com/bperucchi/moip-ruby.git"
 gem "pag_seguro", "0.1.18", git: "git://github.com/bperucchi/pagseguro.git" 
 # gem "pag_seguro", "0.1.18", path: "../../2 - plugins/pagseguro" 
 gem "pagamento_digital", git: "git://github.com/bperucchi/pagamento_digital.git"
 gem 'state_machine', '1.2.0'
-gem 'ancestry', '1.3.0'
+gem 'ancestry', '2.1.0'
 
 gem "paperclip", :git => "git://github.com/thoughtbot/paperclip.git"
 
@@ -59,15 +63,15 @@ gem 'imentore-admin', path: '../imentore-admin'
 
 # Gems used only for assets and not required
 # in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.5'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'uglifier', '>= 1.0.3'
-end
+# group :assets do
+  gem 'sass-rails',   '>= 4.0.0'
+  gem 'coffee-rails', '>= 4.0.0'
+  gem 'uglifier', '>= 1.3.0'
+# end
 
-gem 'jquery-rails', '3.1.2'
-gem 'jquery-datatables-rails', git: 'git://github.com/rweng/jquery-datatables-rails.git'
-gem 'jquery-ui-rails', '5.0.0'
+gem 'jquery-rails', '3.1.3'
+gem 'jquery-datatables-rails', '3.3.0'
+gem 'jquery-ui-rails', '5.0.5'
 ## I don't know why I was use lately version
 # gem 'jquery-datatables-rails', '~> 2.1.10.0.3'
 
@@ -83,6 +87,11 @@ group :development do
   gem 'capistrano-rails', '~> 1.1.1'
   gem 'capistrano-bundler', '~> 1.1.3'
   gem 'capistrano-rbenv', '~> 2.0.2'
+
+  gem 'better_errors'
+  gem 'binding_of_caller'
+  gem 'meta_request'
+
 end
 
 # To use debugger
