@@ -1,6 +1,6 @@
 # module Imentore
 #   class SessionsController < Devise::SessionsController
-#     skip_before_filter :require_no_authentication
+#     skip_before_action :require_no_authentication
 
 
 #     # def new
