@@ -3,11 +3,11 @@ module Imentore
     belongs_to :store
     belongs_to :admin_theme, :class_name => AdminImentore::Theme, :foreign_key => "admin_imentore_theme_id"
     
-    has_many :templates, :dependent => :destroy
+    has_many :templates, class_name: Imentore::Template, :dependent => :destroy
     has_many :assets, :dependent => :destroy
 
 
-    scope :theme_system, where(system: true)
+    scope :theme_system, -> { where(system: true) }
     # scope :default, where(default: true)
 
     def default_layout
