@@ -1,8 +1,8 @@
 module Imentore
   module Client
     class DashboardController < Client::BaseController
-      skip_before_filter :check_store, :authorize_client, :current_cart, :only=>[:key_session]
-      skip_before_filter :verify_authenticity_token#, :if => Proc.new { |c| c.request.format == 'application/json' }
+      skip_before_action :check_store, :authorize_client, :current_cart, :only=>[:key_session]
+      skip_before_action :verify_authenticity_token#, :if => Proc.new { |c| c.request.format == 'application/json' }
       respond_to :json
 
       def index
