@@ -1,7 +1,7 @@
 module Imentore
   class CouponsController < BaseController
 
-    skip_before_filter :verify_authenticity_token, :if => Proc.new { |c| c.request.format == 'application/json' }
+    skip_before_action :verify_authenticity_token, :if => Proc.new { |c| c.request.format == 'application/json' }
     respond_to :json, only: [:add_coupon]
 
     def add_coupon
