@@ -15,9 +15,9 @@ module Imentore
     has_many :categories, :through => :categories_products, :source => :category
     has_many :feedbacks, as: :feedbackable, dependent: :destroy
 
-    scope :active, where(active: true)
-    scope :featured, where(featured: true)
-    scope :product_search, lambda { |search| { :conditions => ["name like ?", "%#{search}%"] } }
+    scope :active,         -> { where(active: true) }
+    scope :featured,       -> { where(featured: true) }
+    scope :product_search, -> (object){ where("name like ?", "%#{search}%") }
 
     validates :name, :handle, :store, presence: true
     validates :handle, uniqueness: { scope: :store_id }
