@@ -5,8 +5,8 @@ module AdminImentore
       "admin_imentore_templates"
     end
 
-    scope :layouts, where(kind: 'layout')
-    scope :templates, where(kind: 'template')
+    scope :layouts,   -> { where(kind: 'layout') }
+    scope :templates, -> { where(kind: 'template') }
 
     belongs_to :theme
     has_many :stores_templates, :class_name => Imentore::Template, :foreign_key => "admin_imentore_template_id"
