@@ -14,22 +14,13 @@ module Old
    :username => 'imentore2',
    :password => '123123'
    )
-
- has_many :images,
-   :class_name => Old::Image,
-   :as => :imageable,
-   :dependent => :destroy,
-   :finder_sql => proc {" SELECT imentore_images.* FROM imentore_images  WHERE imentore_images.imageable_id = #{self.id} AND imentore_images.imageable_type = 'Imentore::ProductVariant' "}
-
   
-  # self.abstract_class = true
-  #  establish_connection(
-  #  :adapter  => 'mysql2',
-  #  :database => 'go2b_production',
-  #  :host     => 'app.imentore.com.br',
-  #  :username => 'imentoreapp',
-  #  :password => 'app0p..za'
-  #  )
+  ## TODO
+  # has_many :images,
+  #   :class_name => Old::Image,
+  #   :as => :imageable,
+  #   :dependent => :destroy,
+  #   :finder_sql => proc {" SELECT imentore_images.* FROM imentore_images  WHERE imentore_images.imageable_id = #{self.id} AND imentore_images.imageable_type = 'Imentore::ProductVariant' "}  
       
   belongs_to :product
 
@@ -38,9 +29,7 @@ module Old
     :foreign_key => :variant_id,
     :dependent => :destroy
 
-
-  scope :not_deleted, where(deleted_at: nil)
-
+  scope :not_deleted, -> { where(deleted_at: nil) }
 
   end
 end
