@@ -10,8 +10,10 @@ ImentoreBase::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  ## RAILS 4
+    # Log error messages when you accidentally call methods on nil.
+    # config.whiny_nils = true
+  config.eager_load = false
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -23,18 +25,19 @@ ImentoreBase::Application.configure do
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :debug
 
-  # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
+  ## RAILS 4
+    # # Only use best-standards-support built into browsers
+    # config.action_dispatch.best_standards_support = :builtin
 
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
+    # # Raise exception on mass assignment protection for Active Record models
+    # config.active_record.mass_assignment_sanitizer = :strict
 
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
+    # # Log the query plan for queries taking more than this (works
+    # # with SQLite, MySQL, and PostgreSQL)
+    # config.active_record.auto_explain_threshold_in_seconds = 0.5
 
-  # Do not compress assets
-  config.assets.compress = false
+    # # Do not compress assets
+    # config.assets.compress = false
 
   # Expands the lines which load the assets
   config.assets.debug = false
