@@ -7,6 +7,9 @@ ImentoreBase::Application.configure do
   # Code is not reloaded between requests
   config.cache_classes = true
 
+  ## RAILS 4
+  config.eager_load = true
+
   # Full error reports are disabled and caching is turned on
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
@@ -14,8 +17,10 @@ ImentoreBase::Application.configure do
   # Disable Rails's static asset server (Apache or nginx will already do this)
   config.serve_static_assets = false
 
-  # Compress JavaScripts and CSS
-  config.assets.compress = true
+  ## RAILS 4
+    # Compress JavaScripts and CSS
+    # config.assets.compress = true
+  config.assets.js_compressor = :uglifier
 
   # Don't fallback to assets pipeline if a precompiled asset is missed
   config.assets.compile = true
