@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-ImentoreBase::Application.config.secret_token = '7854e9c798d42a8c64262b4c9ac99016eb953987c6bfd142ef06149cf7ebf8ee986fd14ff0fb8aaa9f413debfe3098477c165c4d37fecad907548649d43fef8e'
+ImentoreBase::Application.config.secret_key_base = '7823254e9c798d42a8c64262b4c9ac99016eb953987c6bfd142ef06149cf7ebf8ee986fd14ff0fb8aaa9f413debfe3098477c165c4d37fecad907548649d43fef8e'
