@@ -3,8 +3,8 @@ module Imentore
 
     include AdminImentore
 
-    scope :layouts, where(kind: 'layout')
-    scope :templates, where(kind: 'template')
+    scope :layouts,   -> { where(kind: 'layout') }
+    scope :templates, -> { where(kind: 'template') }
     belongs_to :theme
     belongs_to :layout, :class_name => Imentore::Template, :foreign_key => "layout_id"
     belongs_to :admin_template, :class_name => AdminImentore::Template, :foreign_key => "admin_imentore_template_id"
