@@ -1,6 +1,7 @@
 # encoding: utf-8
 require 'active_support/core_ext/big_decimal/conversions'
-require 'active_support/core_ext/float/rounding'
+# require 'active_support/core_ext/float/rounding'
+require 'active_support/number_helper'
 require 'active_support/core_ext/object/blank'
 require 'active_support/core_ext/string/output_safety'
 
