@@ -11,6 +11,6 @@ Imentore::Core::Engine.routes.draw do
   namespace :client do
     resources "address"
     root to: "dashboard#index", as: 'dashboard'
-    match 'key_session',      to: 'dashboard#key_session',        as: 'dashboard_session'
+    get 'key_session',      to: 'dashboard#key_session',        as: 'dashboard_session'
   end
 end
