@@ -5,7 +5,7 @@ module Imentore
       respond_to :json, only: [:index]
 
       def index
-        build_resource
+        @category = begin_of_association_chain.categories.new
         index! do |format|
           format.json do
             @product = current_store.products.find_by_id(params[:product_id])
@@ -22,7 +22,7 @@ module Imentore
       def create
         respond_to do |wants|
           wants.html do
-            @category = current_store.categories.new(params[:category])
+            @category = current_store.categories.new(permitted_params)
             @category.parent = current_store.categories.find_by_id(params[:category][:ancestry])
             if @category.save
               redirect_to admin_categories_path
@@ -47,6 +47,10 @@ module Imentore
       end
 
       protected
+
+      def permitted_params
+        params.require(:category).permit(:name, :handle, :ancestry)
+      end
 
       def begin_of_association_chain
         current_store
