@@ -4,7 +4,7 @@ module Imentore
     # belongs_to :order
     # belongs_to :cart
 
-    scope :active, where(active: true)
+    scope :active, -> { where(active: true) }
     has_many :coupons_orders
     has_many :coupons, :through => :coupons_orders, :source => :coupon
 
